checks: give the changelog listing its own file
checks.go held both the parent command and the whole changelog listing,
while the validate listing already had checks_validate.go. Splitting it
leaves the three files with one job each and the two listings shaped
identically, so a third rule set would have an obvious template.

  checks.go            the parent command, and what both listings share:
                       --format, --severity, the severity filter, and the
                       unknown-subcommand handling
  checks_changelog.go  const, command, flags, runner, output
  checks_validate.go   the same five, in the same order

Names follow: runChecks became runChecksChangelog, matching
runChecksValidate, and outputChecks became outputChangelogRules.
outputValidateRuleIDs became outputValidateRules, since it stopped being
ids alone when it gained descriptions and levels. addChecksValidateFlags
moves next to the command it serves.

Also fixes an error identifier: a failed print in the changelog listing
reported "checks <format>" rather than "checks changelog <format>", the
only place still naming the parent.

Pure moves and renames, no behaviour change: 506 changelog rules and 97
validate rules before and after.
