+++ internal/checks.go
@@ -2,16 +2,12 @@ package internal
 
 import (
 	"fmt"
-	"io"
 	"slices"
 
 	"github.com/oasdiff/oasdiff/checker"
-	"github.com/oasdiff/oasdiff/checker/localizations"
 	"github.com/oasdiff/oasdiff/formatters"
 	"github.com/spf13/cobra"
 )
-
-const checksChangelogCmd = "checks changelog"
 
 // getChecksCmd groups the per-rule-set listings and lists no rules itself:
 // naming one rule set by omission stopped being tenable once `checks validate`
@@ -37,27 +33,6 @@ func getChecksCmd() *cobra.Command {
 	}
 
 	cmd.AddCommand(getChecksChangelogCmd(), getChecksValidateCmd())
-
-	return &cmd
-}
-
-// getChecksChangelogCmd lists the breaking-change and changelog rules, the
-// counterpart of `checks validate`.
-func getChecksChangelogCmd() *cobra.Command {
-
-	cmd := cobra.Command{
-		Use:               "changelog",
-		Short:             "Display changelog and breaking-change checks",
-		Long:              `Display a list of all supported changelog and breaking-change checks.`,
-		Args:              cobra.NoArgs,
-		ValidArgsFunction: cobra.NoFileCompletions,
-		RunE:              getRun(runChecks),
-	}
-
-	// Registered per command rather than inherited: viper binds a command's own
-	// persistent flags (see bindFlags), so an inherited flag would parse but
-	// never reach the config.
-	addChecksChangelogFlags(&cmd)
 
 	return &cmd
 }
@@ -117,25 +92,6 @@ func noSubcommandArgs(cmd *cobra.Command, args []string) error {
 	return err
 }
 
-// addChecksChangelogFlags registers the flags for the changelog listing.
-// --lang belongs here and not on the validate listing: these descriptions are
-// localized, and --tags likewise, since only these rules carry tags.
-func addChecksChangelogFlags(cmd *cobra.Command) {
-	addChecksFormatFlags(cmd)
-	addChecksSeverityFlag(cmd)
-	enumWithOptions(cmd, newEnumSliceValue(getAllTags(), nil), "tags", "t", "include only checks with all specified tags")
-	enumWithOptions(cmd, newEnumValue(localizations.GetSupportedLanguages(), localizations.LangDefault), "lang", "l", "language for localized output")
-}
-
-// addChecksValidateFlags registers the flags for the validate listing: output
-// and severity only. The descriptions are plain English rather than localized
-// (a finding's text comes from the parser at runtime), so --lang would have
-// nothing to translate, and validate rules carry no tags.
-func addChecksValidateFlags(cmd *cobra.Command) {
-	addChecksFormatFlags(cmd)
-	addChecksSeverityFlag(cmd)
-}
-
 // addChecksFormatFlags registers the output format flag every listing needs.
 func addChecksFormatFlags(cmd *cobra.Command) {
 	enumWithOptions(cmd, newEnumValue(formatters.SupportedFormatsByContentType(formatters.OutputChecks), string(formatters.FormatText)), "format", "f", "output format")
@@ -163,66 +119,4 @@ func matchSeverity(severity []string, level checker.Level) bool {
 		return slices.Contains(severity, "info")
 	}
 	return true
-}
-
-func runChecks(flags *Flags, stdout io.Writer) (bool, *ReturnError) {
-	return false, outputChecks(stdout, flags, checker.GetAllRules())
-}
-
-func outputChecks(stdout io.Writer, flags *Flags, rules []checker.BackwardCompatibilityRule) *ReturnError {
-
-	format := flags.getFormat()
-
-	// formatter lookup
-	formatter, err := formatters.Lookup(format, formatters.FormatterOpts{
-		Language: flags.getLang(),
-	})
-	if err != nil {
-		return getErrUnsupportedFormat(format, checksChangelogCmd)
-	}
-
-	localizer := checker.NewLocalizer(flags.getLang())
-
-	// filter rules
-	severity := flags.getSeverity()
-	checks := make(formatters.Checks, 0, len(rules))
-	for _, rule := range rules {
-		if !matchSeverity(severity, rule.Level) {
-			continue
-		}
-
-		// tags
-		if !matchTags(flags.getTags(), rule) {
-			continue
-		}
-
-		commentKey := rule.Id + "-comment"
-		mitigation := localizer(commentKey)
-		if mitigation == commentKey {
-			mitigation = ""
-		}
-
-		checks = append(checks, formatters.Check{
-			Id:          rule.Id,
-			Level:       rule.Level.String(),
-			Direction:   rule.Direction.String(),
-			Area:        rule.Area.String(),
-			Kind:        rule.Kind.String(),
-			Action:      rule.Action.String(),
-			Description: localizer(rule.Description),
-			Mitigation:  mitigation,
-		})
-	}
-
-	// render
-	slices.SortFunc(checks, checks.SortFunc)
-	bytes, err := formatter.RenderChecks(checks, formatters.NewRenderOpts())
-	if err != nil {
-		return getErrFailedPrint("checks "+format, err)
-	}
-
-	// print output
-	_, _ = fmt.Fprintf(stdout, "%s\n", bytes)
-
-	return nil
 }
