+++ internal/checks_test.go
@@ -85,15 +85,3 @@ func Test_ChecksSubcommandsListTheirRules(t *testing.T) {
 		})
 	}
 }
-
-// The listing flags moved to the subcommand, so they have to work there.
-func Test_ChecksChangelogAcceptsTheListingFlags(t *testing.T) {
-	var stdout bytes.Buffer
-	require.Zero(t, internal.Run(
-		cmdToArgs("oasdiff checks changelog -l ru --tags decrease,parameters --severity info,warn,error --format json"),
-		&stdout, io.Discard))
-
-	var checks []map[string]any
-	require.NoError(t, json.Unmarshal(stdout.Bytes(), &checks))
-	require.NotEmpty(t, checks)
-}
